+++ Package.swift
@@ -5,7 +5,7 @@ import PackageDescription
 
 let package = Package(
     name: "ThemeKit",
-    platforms: [.iOS(.v17)],
+    platforms: [.iOS(.v17), .macOS(.v14)],
     products: [
         .library(name: "ThemeKit",        targets: ["ThemeKit"]),
         .library(name: "ThemeKitSwiftUI", targets: ["ThemeKitSwiftUI"]),