+++ Sources/ThemeKit/SystemColorScheme.swift
@@ -5,7 +5,9 @@
 //  Created by Ademola on 07/06/2026.
 //
 
+#if canImport(UIKit)
 import UIKit
+#endif
 
 /// A `Codable` representation of the system light/dark color scheme.
 ///
@@ -19,6 +21,7 @@ public enum SystemColorScheme: Int, Codable, Sendable {
   /// Dark appearance.
   case dark = 2
 
+#if canImport(UIKit)
   /// The corresponding `UIUserInterfaceStyle` for this scheme.
   public var uiUserInterfaceStyle: UIUserInterfaceStyle {
     UIUserInterfaceStyle(rawValue: rawValue) ?? .unspecified
@@ -28,4 +31,5 @@ public enum SystemColorScheme: Int, Codable, Sendable {
   public init(_ style: UIUserInterfaceStyle) {
     self = SystemColorScheme(rawValue: style.rawValue) ?? .unspecified
   }
+#endif
 }