+++ Sources/ThemeKitSwiftUI/Extensions/UIColor+Color.swift
@@ -5,9 +5,11 @@
 //  Created by Ademola on 07/06/2026.
 //
 
+#if canImport(UIKit)
 import SwiftUI
 
 /// Bridges `UIColor` to SwiftUI's `Color`.
 extension UIColor {
   public var color: Color { Color(uiColor: self) }
 }
+#endif