+++ Sources/ThemeKitUIKit/Extensions/UITraitChangeObservable+Combine.swift
@@ -6,6 +6,7 @@
 //
 //  From: https://gist.github.com/pookjw/c32ca552d962ad17cc56fd74f6ad6abf
 
+#if canImport(UIKit)
 import Combine
 import UIKit
 
@@ -28,3 +29,4 @@ extension UITraitChangeObservable where Self: AnyObject {
       })
   }
 }
+#endif